backend/requirements-api.txt
# ============================================
# Phase 3: Production FastAPI Backend
# ============================================
# For the REST API server that communicates with Ollama

# Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0

# Request Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client for Ollama Communication
httpx==0.26.0

# Optional: Database Support (SQLite)
sqlalchemy==2.0.25
alembic==1.13.1

# File Upload Support
python-multipart==0.0.6

# CORS & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Data Processing
pandas==2.1.4
numpy==1.26.2
